Experimental: PKCS#11 Support for Multiple Applications on One Card
The Yubikey is the first card where two different applications are supported
by different card-*.c and pkcs15-*.c drivers. OpenSC in the past has assumed
that there would be only one card driver per card. pkcs11/slot.c is modified
to allow for each slot to have a different card driver.

This is just a proof of concept, because:

It only selects the Yubikey and assumes the only applications that could
be on the card are PIV-II and OpenPGP. This needs to be generalized or added
to the opensc.conf.

There may be locking issues. Each card driver has its only card_lock, but share
the reader-pcsc.c lock. A multi-threaded application might have sessions
open to both drivers and there could be a dead lock situation. (not easy to test.)

It is not well tested.

 On branch pkcs11-multi-application
 Changes to be committed:
	modified:   libopensc/card-openpgp.c
	modified:   libopensc/card-piv.c
	modified:   pkcs11/slot.c