Better error debug messages for Minidriver (first cut)
Add MD_FUNC_CALLED(pCardData, level) and  MD_FUNC_RETURN(pCardData, level, ...)
macros.

Handles are type __int3264 in VS2015 are casted as size_t when printing so
all bytes are printed. size_t on Windows are also treated as 32 or 64 bits.
SC_FORMAT_LEN_SIZE is used in the format.
 (Works with VS2105 needs to be tested on other platforms.

 On branch minidriver-4
 Changes to be committed:
	modified:   src/minidriver/minidriver.c

 Changes not staged for commit:
	modified:   Makefile.mak
	modified:   etc/opensc.conf
	modified:   src/Makefile.mak
	modified:   win32/Make.rules.mak
	modified:   win32/Makefile.mak