Fix SM handling of command chaining for short apdus
Allow driver to tell apdu.c to pass to sm_get_apdus to have it do
the chaining. Without this change, apdu.c will do the command chaining
and pass each chained apdu to SM. This may be OK for some,
but not PIV which uses short APDUs and encryption and MAC for data
before doing the chaining.

The driver sets SC_APDU_FLAGS_SM_CHAINING flag and SC_APDU_FLAGS_CHAINING.
which causes apdu.c to pass the plain adpu to card->sm_ctx.ops.get_sm_apdu.

Without the SC_APDU_FLAGS_SM_CHAINING the flow of SM is not changed.

 On branch PIV-4-extensions
 Changes to be committed:
	modified:   apdu.c
	modified:   types.h