First cut at genkey from pkcs11
PKCS11 3.0 alloes the curvename as printable string to be
sent inplace of OID.  These are mapped into OpenSC known
curve names, OID and size.

in final version this comment will be squashed.

Got as far using openpgp but fails on  non-unique-ID
pkcs11-tool also does not know how to delete a key.
Looks like changes in pkcs15init needs to be made  for
this as gpg-card says "Encryption key....: [none]"
indicating it could be created.

 On branch X25519-improvements
 Changes to be committed:
	modified:   libopensc/card-openpgp.c
	modified:   libopensc/opensc.h
	modified:   libopensc/pkcs15-pubkey.c
	modified:   libopensc/pkcs15.h
	modified:   pkcs11/framework-pkcs15.c
	modified:   pkcs15init/pkcs15-lib.c