card-piv.c - Use NIST_SM_FLAGS_SM_CLOSE_ACCEPT_ERRORS
The flag is used by reader_lock_obtained when SM should still
be active. It uses a SM encoded VERIFY pin command to get tries left.

The command will only responded correctly when:
-  The selected applet has not changed
-  The card and application SM have same shared secret
-  and the Mac counters are the same

If some other process has accessed the card and set up its own
SM  or selected a different applet the shared secret would be
different as the application uses an ephemeral EC key and the
card generates a random nonce. Sm and Login state are reset
in PIV card if a different applet is selected.

PIV allows for some non SM protected operation to be done
to a card even when SM is active such as get data and and verify.
This is the original way PIV cards worked before sp800-74-4.
So this still allows for multiple processes to use the same
the card when the first process has verified the pin.

If SM was not  not active, fall back to the SELECT AID
and to a VERIFY for login state.

Note: the SM encoded VERIFY also returns the login state.

This code has been tested by running `pkcs11-tool --test --login --pin xxxxx`
in teo processes both using SM  and watching  the inter actions
which include having to reestablish SM authenticaion.

 On branch sm-nist
 Changes to be committed:
	modified:   src/libopensc/card-piv.c