card-myeid.c - Problems with select_aid from reader_lock_obtained
Select_aid when done between select_file and read_binary appears to clear
the selected file on the card causing the read_binary to fail.
The select_file and read_binary are each treated as separate SCard
transactions rather then being in the same transaction.
Since a select_aid is a special case of select_file, it appears
the card treats them the same.

Reader_lock_obtained is then called before both. One of the
functions of reader_lock_obtained it to make sure the correct applet
is selected on cards that have multiple applets. Myeid card now
has both a myeid and piv applet that could be called by different
processes using a different applet on the card.

This commit turns off the select_aid call and also does not use SM
unless env MYEID_USE_SM is defined. MYEID_USE_SM can be set to 'a' (always),
'n' (never) or any other value which to use SM only on verify
and crypto. This is a temp fix which needs to be added to opensc.conf.
For now, MYEID_USE_SM can be used for testing.

The the issue of select_aid not being treated specially, can be
addressed with an extra call to sc_lock before a select_file and
and an extra call to sc_unlock after the last read_binary.

(Other cards may have the same issue with select_aid and select_file
but do not address the multiple applets problem.)

 On branch myeid-4.9.1
 Changes to be committed:
	modified:   card-myeid.c