pkcs15-cert.c - Add support of CVC type certificates
X509 and CVC certificates are parsed differently than X509.
X509 ASN.1 starts with SEQUENCE, CVC start with 0x7F21.

Not all CVC certificates are the same. The intent of these
changes are to treat a CVC as DER, so user can read a
CVC from the card and write a CVC to a card without having
to parse it. Card drivers, tools or SM routines can continue
to encode or decode their CVCs as needed.

Within pkcs15 routines certificate structures now have a "type"
which is read from the CDF and provide by user when writing
a certificate and updating the CDF using PKCS15 or PKCS11.
The sc_pkcs15_encode_cdf and sc_pkcs15_encode_cdf
are modified to get/put the type in the CDF
The CDF entry for X509 does not change.

There are a few places, where OpenSC tries to get the SPKI from
a certificate using sc_pkcs15_pubkey_from_cert which now looks
at the ASN.1 to determine the type of certificate. If X509
parse_x509_cert is used to get more then just the SPKI.
If CVC, parse_cvc_cert is used and just saves the DER.
(TODO try to get EC OID and ECPoint.)

The sm/sm-nist routines have a sm_nist_decode_cvc for the NIST
format of a CVC. Currently the CVC is created ouside of OpenSC.