pkcs15init/pkcs15-lib.c - When storing a CVC certificate
When storing a CVC certificate, an ID is required, as there
is no way to use sc_pkcs15init_select_intrinsic_id
The ID is expected to match the ID of the public and private keys.
The ID is also used for PKCS11.

The previous sc_pkcs15_find_object_by_iD used SC_PKCS15_TYPE_CERT
which worked  as certs where covered by SC_PKCS15_TYPE_CERT. We now search
only the same type of certificate.

 On branch myeid-4.9.1
 Changes to be committed:
	modified:   src/pkcs15init/pkcs15-lib.c