pkcs11-tool.c - Support --cert-type option
The default cert-type is x509 for CKC_X_509
the other choice is cvc for CKK_CVC.
(CKK_CVC is an OpenSC vendor provided CKK.
This allows the caller to load any CVC in DER to a card.

Only CKK_X_509 certs can be parsed.
CVC certificates can be read/writen to/from card in only DER format.

 On branch myeid-4.9.1
 Changes to be committed:
	modified:   src/tools/pkcs11-tool.c