pkcs15-piv.c - support loading certificates into retired key slots
If no certificate is present on card, but env PIV_xx_KEY points at
file with public key  in DER format card-piv.c will allow
the key to be used  with out a certificate. The intent is OpenSC
can be used to sign a CSR or self signed certificate which can then be loaded.

where xx is one of the supported keys  i.e. 9A, 9C, 9D, 9E or
for retired keys 82 to 93

 On branch piv-retired-keys
 Changes to be committed:
	modified:   pkcs15-piv.c